Checkear null
git-svn-id: http://jira.uji.es/repos/uji/PAR/uji-par/trunk@16970 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ uji-par-base/src/main/java/es/uji/apps/par/dao/ComprasDAO.java
@@ -92,7 +92,7 @@ public class ComprasDAO
     {
         CompraDTO compra = getCompraById(idCompra);
         
-        if (!compra.getPagada())
+        if (compra!=null && !compra.getPagada())
         {
             entityManager.remove(compra);
         }
